FIxed error, added docs
